+++ NEWS.md
@@ -1,5 +1,6 @@
 # levelnet 0.2.0.9000
 
+* MatMul is now done with RcppEigen (random network sim in sdsm now twice as fast)
 * implemented scobit function in C++
 * added scobit gradient function in C++
 * MLE of scobit model is now performed with `optim` and the gradients